backend/requirements.txt
bs4
flask
flask-cors
flask-restful
requests
langchain
langchain_community
langchain-ollama
langgraph
faiss_cpu
Flask[async]
sentence-transformers
langchain-core
langchain-community
langchain-pinecone
google-generativeai